feat: add record_campaign_completion instruction

+++ programs/cmpgn/src/state/campaign.rs
@@ -5,6 +5,6 @@ use anchor_lang::prelude::*;
 pub struct Campaign {
     pub game_authority: Pubkey,
     pub campaign_id: u8,
-    pub total_completions: u16,
+    pub total_completions: u8,
     pub bump: u8,
 }
